feat: restore kona command, add filters to kona command
